CVertexBuffer: Make use of size_t where applicable
Same behavior, less truncations.

+++ src/Core/OpenGL/CVertexBuffer.cpp
@@ -270,7 +270,7 @@ void CVertexBuffer::SetSkin(CSkin *pSkin)
     mpSkin = pSkin;
 }
 
-uint32 CVertexBuffer::Size() const
+size_t CVertexBuffer::Size() const
 {
     return mPositions.size();
 }
